Restructure codebase with modular organization
